Time-based recency: fade over 24h wall-clock instead of snapshot count
Recency now stores wall-time of last activity instead of a counter.
Nuclei fade over 24 real hours and evict after 48h. This gives
consistent visibility regardless of snapshot frequency — no more
dropping to ~100 concepts when snapshots arrive in bursts.

+++ viewer/src/main.zig
@@ -26,7 +26,7 @@ pub fn main() !void {
     std.debug.print("Loaded {d} positions\n", .{nd.positions.count()});
 
     // --- Spawn worker thread (handles all snapshot loading + live) ---
-    var running_recency = std.StringHashMap(u32).init(allocator);
+    var running_recency = std.StringHashMap(i64).init(allocator);
     var nds = live.NdSnapshot.init(&nd, &running_recency);
     var queue = live.LiveQueue{};
     queue.spawn(&nds);
